analytics :    - remove event and property files    - add ExceptionHandler    - modify some fields in jpa and domain    - improve process

+++ analytics/adapter/infra/analytics-rdb/src/main/kotlin/ir/smartech/cro/analytics/rdb/adapter/FunnelRepositoryAdapter.kt
@@ -1,34 +1,39 @@
 package ir.smartech.cro.analytics.rdb.adapter
 
+import ir.smartech.cro.analytics.domain.common.api.utils.ErrorCodes
+import ir.smartech.cro.analytics.domain.common.api.utils.ResponseException
 import ir.smartech.cro.analytics.rdb.mapper.funnel.FunnelMapper
 import ir.smartech.cro.analytics.rdb.repository.JpaFunnelRepository
 import ir.smartech.cro.analytics.domain.funnel.api.entity.Funnel
 import ir.smartech.cro.analytics.domain.funnel.spi.FunnelRepository
-import org.springframework.data.crossstore.ChangeSetPersister.NotFoundException
 import org.springframework.stereotype.Component
 import org.springframework.transaction.annotation.Transactional
-import java.util.*
 
 @Component
+@Transactional
 class FunnelRepositoryAdapter(private val repo: JpaFunnelRepository, private val funnelMapper: FunnelMapper) :
     FunnelRepository {
-    override fun save(entity: Funnel): Funnel? {
+    override fun save(entity: Funnel?): Funnel? {
+        if (entity == null) return null
         // map to jpa entity
-        val jpaFunnel = funnelMapper.toDestination(entity) ?: throw NotFoundException()
+        val jpaFunnel = funnelMapper.toDestination(entity)
         val saved = repo.save(jpaFunnel)
         // map to domain entity
         return funnelMapper.toSource(saved)
     }
 
-    @Transactional
-    override fun findById(id: Int): Optional<Funnel> {
-        val dbObject = repo.findById(id).get()
-        return Optional.of(funnelMapper.toSource(dbObject)!!)
+    override fun findById(id: Int): Funnel? {
+        val dbObject = repo.findById(id)
+            .orElseThrow { ResponseException(ErrorCodes.NOT_FOUND, "the funnel with $id id not found") }
+        return funnelMapper.toSource(dbObject)
     }
 
-    override fun findByIdAndProjectId(id: Int, projectId: Int): Optional<Funnel> {
-        val dbObject = repo.findByIdAndProjectId(id, projectId)
-        return Optional.of(funnelMapper.toSource(dbObject)!!)
+    override fun findByIdAndProjectId(id: Int, projectId: Int): Funnel? {
+        val dbObject = repo.findByIdAndProjectId(id, projectId) ?: throw ResponseException(
+            ErrorCodes.NOT_FOUND,
+            "the funnel with $id id not found"
+        )
+        return funnelMapper.toSource(dbObject)
     }
 
     @Transactional
@@ -46,8 +51,9 @@ class FunnelRepositoryAdapter(private val repo: JpaFunnelRepository, private val
         repo.deleteById(id)
     }
 
-    override fun delete(entity: Funnel) {
-        val jpaFunnel = funnelMapper.toDestination(entity) ?: throw NotFoundException()
+    override fun delete(entity: Funnel?) {
+        if (entity == null) return
+        val jpaFunnel = funnelMapper.toDestination(entity)
         repo.delete(jpaFunnel)
     }
 }
